Homogenisation of protocols all over the place, removed nonlinear (finally!), change names to accomodate to the src structure
