+++ cloelite/__init__.py
@@ -2,5 +2,4 @@ __all__ = []
 
 from cloelite.cosmology import *
 from cloelite.likelihood import *
-from cloelite.nonlinear import *
 from cloelite.observables import *
