+++ cloelite/summary_statistics/__init__.py
@@ -1,2 +1,3 @@
-from cloelite.summary_statistics.two_point import *
+from cloelite.summary_statistics.angular_two_point import *
+from cloelite.summary_statistics.legendre_multiples import *
