+++ cloelite/summary_statistics/legendre_multiples.py
@@ -1,78 +1,300 @@
 # cloelite imports
-from cloelite.cosmology.cosmology import Background, Perturbations
+from cloelite.observables.spectro import SpectroPower
+from cloelite.auxiliary.math_utils import legendre, simps_jax # I made it auto-diff :) 
 
 # General imports
 from typing import Protocol, Union, TypeVar, Optional, Generic
 import numpy as np  # type: ignore
-import jax.numpy as jnp
-
-T = TypeVar("T", bound=Union[jnp.ndarray, np.ndarray])
+import jax.numpy as jnp # type: ignore
+from scipy import integrate # type: ignore
+from scipy.special import roots_legendre # type: ignore
 
 """
 
 ## Notes:
 
-- Legendre multiples protocol to implement different spectro external codes
+- Class to calculate the legendre multiples given a SpectroPower instance
 
 """
 
-class LegendreMultipoles(Protocol, Generic[T]):
-    @property
-    def linear_perturbations(self) -> Perturbations:
-        """
-        Stores perturbations obj
-        """
-        ...
+class LegendreMultipoles: #this class depends on SpectroPower
+    def __init__(self, spectro_power: SpectroPower,
+                 nbar: float, sigmaz: float, fout: float, redshift: float):
+        self.spectro_power = spectro_power
+        mu_min = 0.0
+        mu_max = 1.0
+        mu_samp = 101
+        self.mu_grid = np.linspace(mu_min, mu_max, mu_samp)
+        self.nbar = nbar
+        self.sigmaz = sigmaz
+        self.fout = fout
+        self.redshift = redshift
 
-    @property
-    def background_fiducial(self) -> Background:
+    def _q_AP_tr(self, zs: np.ndarray) -> np.ndarray:
+        r"""AP distortion parameter transversal to the line of sight
+        .. math::
+            q_{\perp}(z) &= \frac{D_{\rm M}(z)}{D_{\rm M,fid}(z)}\\
+        Parameters
+        ----------
+        z: np.ndarray
+           Redshift
+        Returns
+        -------
+        q_tr: np.ndarray
+           Transversal AP parameter
         """
-        Stores background obj
-        """
-        ...
+        # as long as SpectroPower has the attribute background and background_fiducial
+        # this will always work, because it is dependent on Background protocol that will
+        # always have angular_diameter_distance
+        return (self.spectro_power.background.angular_diameter_distance(zs)
+                / self.spectro_power.background_fiducial.angular_diameter_distance(zs))
 
-    @property
-    def redshifts(self) -> T:
+    def _q_AP_lo(self, zs: np.ndarray) -> np.ndarray:
+        r"""AP distortion parameter parallel to the line of sight
+        .. math::
+            q_{\parallel}(z) &= \frac{H_{\rm fid}(z)}{H(z)}\\
+        Parameters
+        ----------
+        z: np.ndarray
+           Redshift
+        Returns
+        -------
+        q_tr: np.ndarray
+           Parallel AP parameter
         """
-        Set mean redshifts bins
-        """
-        ...
-    
-    @property
-    def sigmaz(self) -> T:
-        """
-        Set redshifts bin error
-        """
-        ...
+        return (self.spectro_power.background_fiducial.hubble_parameter(zs)
+                /self.spectro_power.background.hubble_parameter(zs))
 
-    @property
-    def fout(self) -> T:
-        """
-        Set redshifts bin error
-        """
-        ...
+    def _ensure_array(self, param):
+        if np.isscalar(param):
+            param = np.array([param])
+        return np.asarray(param)
 
-    @property
-    def nbar(self) -> T:
+    def _k_AP(self, k: np.ndarray, mu: np.ndarray, zs: float,
+              use_AP: Optional[bool] = True) -> np.ndarray:
+        r"""AP-distorted wavenumber
+        .. math::
+            k(k_{\rm fid},\mu_{\rm fid}, z) &= k_{\rm fid} \
+            \left[\frac{(\mu_{\rm fid})^2}{q_\parallel^2(z)} + \
+            \frac{1-(\mu_{\rm fid}^2)}{q_\perp^2(z)}\right]^{1/2}
+        Parameters
+        ----------
+        k: np.ndarray
+           Fiducial wavenumber
+        mu: np.ndarray
+           Fiducial angle (cosinus) to the line of sight
+        z: float
+           Redshift
+        use_AP: bool
+            Flag to switch between with and without AP corrections
+        Returns
+        -------
+        kAP: np.ndarray
+           AP-distorted wavenumber
         """
-        Set redshifts bin error
-        """
-        ...
+        q_tr = self._q_AP_tr(zs) if use_AP else 1.0
+        q_lo = self._q_AP_lo(zs) if use_AP else 1.0
+        return np.outer(k, np.sqrt(mu**2 / q_lo**2 + (1.0-mu**2) / q_tr**2))
 
-    def power_multipoles(self, k: T,
-                         ells: T, **args) -> T:
+    def _mu_AP(self, mu: np.ndarray, zs: float,
+               use_AP: Optional[bool] = True) -> np.ndarray:
+        r"""AP-distorted angle (cosinus) to the line of sight
+        .. math::
+            \mu(\mu_{\rm fid}, z) &= \frac{\mu_{\rm fid}}{q_\parallel(z)} \
+            \left[\frac{(\mu_{\rm fid})^2}{q_\parallel^2(z)} + \
+            \frac{1-(\mu_{\rm fid}^2)}{q_\perp^2(z)}\right]^{-1/2}
+        Parameters
+        ----------
+        mu: np.ndarray
+           Fiducial angle (cosinus) to the line of sight
+        z: float
+           Redshift
+        use_AP: bool
+            Flag to switch between with and without AP corrections
+        Returns
+        -------
+        muAP: np.ndarray
+           AP-distorted angle (cosinus) to the line of sight
+        """
+        q_tr = self._q_AP_tr(zs) if use_AP else 1.0
+        q_lo = self._q_AP_lo(zs) if use_AP else 1.0
+        return mu / q_lo / np.sqrt(mu**2 / q_lo**2 + (1.0-mu**2) / q_tr**2)
+
+    def Pk2d_noise(self, k: np.ndarray, mu: np.ndarray,
+                    parameters: dict) -> np.ndarray:
+        r"""2D power spectrum from expansion of stochastic field
+        Parameters
+        ----------
+        k: np.ndarray
+            Wavenumber
+        mu: np.ndarray
+            Angle (cosinus) to the line of sight
+        parameters: dict
+            Ensemble of cosmological and nuisance parameters
+        Returns
+        -------
+        Pk2d_noise: np.ndarray
+            2D power spectrum from expansion of stochastic field
+        """
+        noise = parameters['NP0'] + k**2 * (parameters['NP20'] +
+                                            parameters['NP22'] *
+                                            legendre(2, mu))
+        return noise / self.nbar
+
+    def _damping_function(self, k: np.ndarray, mu: np.ndarray,
+                          parameters: dict) -> np.ndarray:
+        r"""Damping function due to GCsp redshift uncertainty
+        Parameters
+        ----------
+        k: np.ndarray
+            Wavenumber
+        mu: np.ndarray
+            Angle (cosinus) to the line of sight
+        parameters: dict
+            Ensemble of cosmological and nuisance parameters
+        Returns
+        -------
+        damping_function: np.ndarray
+            Damping function due to GCsp redshift uncertainty
+        """
+        sigma_z = parameters['sigmaz']
+        sigma_r = 299792.458 * sigma_z / \
+            self.spectro_power.background_fiducial.hubble_parameter(parameters['z'])
+        return np.exp(-k**2 * mu**2 * sigma_r**2)
+
+    def _Pk2d_noise(self, k: np.ndarray, mu: np.ndarray,
+                    parameters: dict) -> np.ndarray:
+        r"""2D power spectrum from expansion of stochastic field
+        Parameters
+        ----------
+        k: np.ndarray
+            Wavenumber
+        mu: np.ndarray
+            Angle (cosinus) to the line of sight
+        parameters: dict
+            Ensemble of cosmological and nuisance parameters
+        Returns
+        -------
+        Pk2d_noise: np.ndarray
+            2D power spectrum from expansion of stochastic field
+        """
+        noise = parameters['NP0'] + k**2 * (parameters['NP20'] +
+                                            parameters['NP22'] *
+                                            legendre(2, mu))
+        return noise / self.nbar
+
+    def _Pk2d_tot(self, k: np.ndarray, mu: np.ndarray,
+                  parameters: dict) -> np.ndarray:
+        r"""Total 2D power spectrum (including RSD, systematics, and noise)
+        Parameters
+        ----------
+        k: np.ndarray
+            Wavenumber
+        mu: np.ndarray
+            Angle (cosinus) to the line of sight
+        parameters: dict
+            Ensemble of cosmological and nuisance parameters
+        Returns
+        -------
+        Pk2d_tot: np.ndarray
+            Total 2D power spectrum (including RSD, systematics, and noise)
+        """
+        # here is where you see how to call the SpectroPower method
+        # if the instance you are passing follows the SpectroPower protocol
+        # this will work
+
+        return (self.spectro_power.Pk2d_rsd(k, mu) *
+                self._damping_function(k, mu, parameters) *
+                (1.0 - parameters['fout'])**2 +
+                self._Pk2d_noise(k, mu, parameters))
+
+    def power_multipoles(self, k: np.ndarray, parameters: dict, #not sure about dicts
+                         ells: Optional[np.ndarray] = None,
+                         use_AP: Optional[bool] = True) -> dict:
         r"""Power spectrum Legendre multipoles
         Parameters
         ----------
-        k: jax.numpy.ndarray or numpy.ndarray
+        k: np.ndarray
             Wavenumber
-        ells: jax.numpy.ndarray or numpy.ndarray
+        parameters: dict
+            Ensemble of cosmological and nuisance parameters
+        ells: np.ndarray
             Legendre multipole order
-        args: EFTofLSS parameters
-
+        use_AP: bool
+            Flag to switch between with and without AP corrections
         Returns
         -------
-        multipoles: jax.numpy.ndarray or numpy.ndarray
+        multipoles: dict
             Power spectrum Legendre multipoles
         """
-        ...
+        #self.update(**parameters) #you don't need update, in reality
+        params = parameters.copy()
+        ells = self._ensure_array(ells) if ells else np.array([0,2,4])
+        AP_factor = (self._q_AP_tr(parameters['z'])**2 *
+                     self._q_AP_lo(parameters['z']) if use_AP else 1.0)
+        prefactors = np.array([(2.0 * m + 1.0) for m in ells]) / 2.0 / \
+            AP_factor
+        multipoles = {}
+        for i,ell in enumerate(ells):
+            multipoles[f'ell{ell}'] = \
+                integrate.simps(self._Pk2d_tot(self._k_AP(k, self.mu_grid,
+                                                          parameters['z'],
+                                                          use_AP=use_AP),
+                                               self._mu_AP(self.mu_grid,
+                                                           parameters['z'],
+                                                           use_AP=use_AP),
+                                               params) *
+                                legendre(ell, self.mu_grid),
+                                self.mu_grid, axis=1)
+            multipoles[f'ell{ell}'] *= (2.0 * prefactors[i])
+        return multipoles
+
+    def convolved_power_multipoles(self, parameters: dict, mixing_matrix=dict):
+        r"""Power spectrum Legendre multipoles convolved with the mixing matrix
+        Parameters
+        ----------
+        parameters: dict
+            Ensemble of cosmological and nuisance parameters
+        Returns
+        -------
+        multipoles_out: dict
+            Convolved power spectrum Legendre multipoles
+        """
+
+        self.mixing_matrix_dict = mixing_matrix
+
+        for key in self.mixing_matrix_dict:
+            self.mixing_matrix_dict[key] = \
+                np.asarray(self.mixing_matrix_dict[key], dtype=np.float32)
+
+        kin0 = self.mixing_matrix_dict['kin0']
+        kin2 = self.mixing_matrix_dict['kin2']
+        kin4 = self.mixing_matrix_dict['kin4']
+        kout = self.mixing_matrix_dict['kout']
+
+        multipoles_in = {}
+        if np.all(kin0 == kin2) and np.all(kin2 == kin4):
+            multipoles_in = self.power_multipoles(k=kin0,
+                                                  parameters=parameters,
+                                                  ells=[0,2,4])
+        else:
+            for ell in [0,2,4]:
+                multipoles_in[f'ell{ell}'] = \
+                    self.power_multipoles(
+                        k=self.mixing_matrix_dict[f'kin{ell}'],
+                        parameters=parameters, ells=[ell])
+
+        for key in multipoles_in:
+            multipoles_in[key] = \
+                np.asarray(multipoles_in[key], dtype=np.float32)
+
+        multipoles_out = {}
+        multipoles_out['k'] = kout
+        for ell in [0, 2, 4]:
+            multipoles_out[f'ell{ell}'] = np.zeros(kout.shape)
+            for ell_prime in [0, 2, 4]:
+                multipoles_out[f'ell{ell}'] += \
+                    np.dot(self.mixing_matrix_dict[f'W{ell}{ell_prime}'],
+                           multipoles_in[f'ell{ell_prime}'])
+
+        return multipoles_out
